Adds `WhereCondition` operators
This adds a number of operators for building and combining
`WhereCondition` values for those that would like to use use them.

In adding these operators I realized we had missed creating binary
boolean functions for `AND` and `OR`. I ended up deciding `whereAnd` and
`whereOr` should simply be those functions and that the existing
versions of the functions were not really necessary. A user can also do
a fold using `whereAnd` or `whereOr` as desired if they have a dynamic
list of conditions. For literals, the operators are available to make
this nice be, such as

```
      (fieldA .== a)
  .&& (fieldB .== b)
  .&& (fieldB .== b)
```

Or, alternatively

```
    whereAnd (fieldA .== a)
  . whereAnd (fieldB .== b)
  $          (fieldB .== b)
```