Adds ability to customize begin transaction sql
In order to test this I ensured that the transaction sql statements made
it into the sql executino callbacks so I could trace them in the test.
This required providing execute functions that both delivery callbacks
and work in IO so that the transaction callback could be made inside the
same `liftIO` as the sql execution.

When adding the test for this I discovered that our `beginTransaction`
expression builder built incorrect sql when used with an isolation
level. I fixed that as well.