Implement a Date parser in attoparsec
The date parser has a small amount of leniency when it comes to
parsing years, the ISO spec states that years should not be less
than 4 digits and the parser allows for smaller years. Postgres
should not give us a date like that, and adding support for that
would mean reimplementing the `decimal` function, so I chose to
allow the leniency.

I added a couple more cases for Date tests.