Adds basic support for custom indices
Custom indices are now supported in Orville, with the following caveats:
* No check is made on the length of the name, however, postgresql has a
  63 character limit
* No check is made to ensure the index name doesn't conflict with
  something else in the schema.

Custom indices can be created by creating an index with a specific name,
to differentiate from regular attribute based indices.

Regular indices are given a default name, so when describe existing
indices during migration, we have to generate a list of migration keys
which includes the Attribute based migration key *and* the Named
migration key. We compare the desired indices against this list of keys
and if any matches, we keep the index, otherwise we drop it.