Adjusts `BooleanExpr` construction to be more re-useable
This changes a number of the Expr functions to take `ValueExpression`
rather than `ColumnName` or `SqlValue` because it allows users to use
them to construct more generic expressions.

The field-definition wrappers for these functions (such as
`fieldEquals`) still have the same type the used to. The implementations
have been adjust accordingly.