Removes wrapper types from SelectOptions fields
This removes the `OrderBy` and `WhereCondition` types which had become
just newtype wrappers around `OrderByExpr` and `BooleanExpr`
respectively. Removing these wrappers allows users to make a smoother
transition from working with `SelectOptions` to building `Expr`
expressions directly because the helpers the are used to using
(`orderByField`, `fieldEquals`, etc) can be still be used.