Removes `Internal` from the `Expr` module
This makes the `Expr` module "officially" public be removing the
`Internal` moniker from the name. To make the documentation page for the
module less intimidating (i.e. not one big page of every single expr
function), I also exposed the modules immediately under `Expr`. In most
cases I inlined small modules such as `OrderByClause` into their parent
to make the docs more cohesive and try to keep the source organization
matching the docs. I made an exception in the case of `Name` because of
the number of distinct modules there and marked the individual name
modules internal to reflect that they are _not_ presented in the docs.