Adds Haddocks to several modules.
Haddocks, primarily module headers, are added to AutoMigration,
EntityTrace, ErrorDetailLevel, and Execution/*. This includes a module
description, which is short, but shows the general pattern for doing
so. I suspect that we will want to expand that pattern going forward.

This is a somewhat arbitrary place to cut this work, and more is to follow.