Adds/updates Haddocks to more modules
Notably, I've made sure the @since is included in everything that I've
gone through thus far, so the docs will show when functionality,
including typeclass instances were added. Thus far everything is since
0.10.0.0, as even if they existed prior, given the major rewrite that
is 0.10.0.0 it seems certain some change was in effect. Additionally
it is much easier, to state them uniformly this way.

This also adds a new helper `unsafeSqlExpression` that eases going
from string literals to something implementing the SqlExpression
class. Previous usages of unsafeFromRawSql combined with fromString
are replaced with this helper to show it's usage.

Still cutting in rather arbitrary places to keep individual reviews
smaller in size.