Add missing GroupBy tests
The tests were also failing because the results were in a different
order. I fixed this by correcting the order, but since the order in SQL
is technically undefined when a sort has not been specified this isn't
the best fix.

However, changing the code to use a Set would also necessitate that I
change `assertEqualSqlRows` too, so I opted to avoid doing that for now.