De-dupes in clause parameters in Plans finding by fields.
PostgreSQL has a limit of 65535 parameters that may be passed with a
query. When complicated Plans are constructed with multiple layers of
`planMany` it's possible to hit this limit because the same value is
passed multiple items within the `planMany` list. This deduplicates the
values passed when building the in clause to avoid the limit in these
cases.