Fixes async exceptions interfering with transaction callback sequences
It was possible for a transaction rollback exception to get delivered to
a caller listening to callbacks if an asynchronous exception occurred
after the begin transaction sql was executed and the open transaction
callback being delivered. In this case the `finally` clause would
trigger and rollback the exception successfully and attempt to deliver
an exception rollback callback. This could be confusing to the caller
because they never received the begin transaction callback. In
particular, the entity trace module raises an error in this scenario.

This avoids the problem be replacing the `finally` call with a call to
our own version of `bracket` (the stock `bracket` is insufficient
because it does not provided enough information to distinguish between
releasing the transaction in success and exception scenarios).