[SC-30704] Adds support for jsonb
Adds a jsonb DataType, Field, and SqlType. No default value was given
for the `jsonbField`.

Tests were added for the jsonb field and SqlType. For the jsonb field
test, I chose to limit the keys and values to alpha-numeric values, despite the
fact that keys and values can be any valid string.