Moves the HDBC based orville-postgresql to be named orville-postgresql-legacy
This makes room to name the libpq version without a suffix. That is
expressly not done at this point so that documentation can be
regenerated. This allows users of the legacy version to be informed
they should update any bookmarks they may have without the
documentation simply changing silently under them.