Adds basic support for concurrent index creation
This adds the option for a user to specify an `IndexDefinition` should
be created asynchronously during migrations.

To do this the migration lock needed to be changed to be acquired at the
session level rather than the transaction level, which means we need to
manually release the lock. I added tests to ensure that this lock
release happens as expected. I also exported the `withTransactionLock`
function both to test it, and to make it available to users in case they
want to use `generateMigrationPlan` and `executeMigrationPlan`
separately.