Haddock cleanup in prep for release
This does a pass through the remaining items to add haddock where it's
missing and standardize @since values and module headers.

We will probably do another pass to add docs to the module headers as
well and the @since may yet change if we decide to release we a
different version number.