Allow users to concurrent index creation in a custom manner
This updates the docs to reflect that PostgreSQL does NOT run concurrent
index creation the background but rather blocks waiting for it to
finish. It also adds a `MigrationOptions` argument to allow the caller
control over which block of migration steps are to be run and what
migration lock to use. This should be enough to allow callers to run the
concurrent indexes in their own user-supplied thread if they desire.