Exposes 'IndexMigrationKey' and related
We had previously exposed these and they were hidden, perhaps by
accident.  As it stands we already have items at the top level that
make use of the type, much like the similar type for constraints, that
users cannot write the signature for themselves or otherwise use the
values provided.