Adds a basic 'OnConflictExpr'
This adds very rudementary support for ON CONFLICT in the
'InsertExpr'. Adding only a type and a helper for 'ON CONFLICT DO
NOTHING'.

The placement of the type alongside the insert related is somewhat
arbitrary compared to being a separate module. It was placed within
the insert because it would appear that postgresql, as of version 16,
only supports it on insert statements so it is not really reusable
elsewhere.

This also changes the 'insertExpr' function to optionally take the new
'OnConflictExpr'. As well as the 'mkInsertExpr' at the
'TableDefinition' level. However, this is not filtered to the highest
levels of the api. The lower levels are changed because the ordering
of the 'ON CONFLICT', must come before the 'RETURNING' part of an
insert and without the change the only way to effectively use the
change would have been to not use the existing expr level
functionality at all. The highest level is unchanged so we think
through how to evolve the api, particularly should we want good
support for 'ON CONFLICT DO SET ...'. This set of low level/high level
feels like a decent tradeoff but revisiting the decisions is without
question, appropriate too.