Adds expr level support for 'TRUNCATE TABLE'
In addition to the 'TruncateTableExpr' a pair of expr helpers for
creating it are added to allow truncation of one or more tables in a
single statement.

Note that this does not touch the automigration code in any way or add
any higher level functionality. Users can now write the faster
TRUNCATE, but execution and management is entirely up to them.