Adds initial aliasing support.
The highlights:

- Moving to use 'Qualified ColumnName' in many/most places. This
allows for the aliasing to be used naturally in those places.

- Add an alias type and function to create a 'Qualified ColumnName'
with it. This signals that 'Alias' is in fact different than
'ColumnName' or 'TableName', and reusing them for aliasing is not
always appropriate.

- Adding a 'MarshallAlias' constructor to the 'SqlMarshaller' so that
an alias can be attached to all of the natural fields in a marshaller
in a single function call from users. Synthetic fields could be
refactored with this facility, but that was left out here to keep the
already big changeset smaller.

- Many of the fieldDefinition helpers now have a version to support
generating various subexpressions with an alias along with the
fieldDefinition. This is a choice to allow the simple case to not deal
with the complexity of the alias, but make it easier for when aliases
are needed.

- A small number of helpers are added to make generation of bits of
SQL, such as a 'TableReferenceList' that includes the 'AS alias' piece.