Adds initial aliasing support.
The highlights:

- Moving to use 'Qualified ColumnName' in many/most places. This
allows for the aliasing to be used naturally in those places.

- Add an alias type and function to create a 'Qualified ColumnName'
with it. This signals that 'Alias' is in fact different than
'ColumnName' or 'TableName', and reusing them for aliasing is not
always appropriate.

- Adding a 'MarshallAlias' constructor to the 'SqlMarshaller' so that
an alias can be attached to all of the natural fields in a marshaller
in a single function call from users. Synthetic fields could be
refactored with this facility, but that was left out here to keep the
already big changeset smaller.

- Adds 'AliasedFieldDefinition' for holding a field that should be
referenced with an alias in SQL.

- Adds a 'SqlComparable' typeclass to allow for comparison in a
polymorphic way. Instances are provided for 'FieldDefinition',
'SyntheticField' and the new 'AliasedFieldDefinition'. Comparison
functions are exposed working with the class. The field definition
functions and operators for comparison are now type restricted
versions of the classy functions. In this way the "simple" case of
FieldDefinitions has the most specific functions, but
'AliasedFieldDefinition' and 'SyntheticField' can both be compared
against.

- A small number of helpers are added to make generation of bits of
SQL, such as a 'TableReferenceList' that includes the 'AS alias' piece.