Adds initial support for triggers and functions.
This adds `Expr` functions for building SQL expressions related to
triggers and functions along with schema level `FunctionDefinition`
and `TriggerDefinition` abstractions.

`FunctionDefinition`s are meant be included directly in `SchemaItem`
lists for auto migration purposes to create the functions that triggers
will call. For the purposes of this change we provide only `mkTriggerFunction`
for construction a new `FunctionDefinition` for a trigger function.

`TriggerDefinition`s are added to a `TableDefinition` via the new
`addTableTriggers` function, similar to indexes and constraints.
For the time being migration of trigger's is done by name only --
existing triggers are never altered. This is due both to the complexity
involved in interpreting the `pg_trigger` catalog entries to determine
whether the trigger needs to be migration and also the fact that
`CREATE OR REPLACE TRIGGER` is not available in PostgreSQL 14 and
would be required for changing most of the properties on triggers.

Co-Authored-By: Paul <paul@flipstone.com>