Adds basic support for PostgreSQL extensions and particularly pg_trgm
At the highest level this does four things:
- Adds support for extensions, including automigration
- Adds a 'IfNotExistsExpr' for use in some extension support.
- Adds support for a subset of functionality provided by the pg_trgm
extension.
- Allows ordering to be done on a synthetic field.

Extension support is added at the expression level, as well as to the
'PgCatalog' tracking infrastructure to support 'AutoMigration'
for loading/unloading extensions.

Support for cascading the load of extensions supported and used
during automigration so that users can be sure to list the extension
they want and any dependencies will be loaded as well.

To be used in conjection with the extension support is some
functionality from pg_trgm. The ability to add both GIN and GIST
indexes are included, as well as overriding the GIST index parameter,
'siglen'. Further, the similarity, word_similarity, and
strict_word_similarity functions are exposed to compare value
expressions. Higher level synthetic fields that compare a column with
a value are also included. This hopefully covers a broad range of
usage such as LIKE queries, column similarity, and the flexibility to
use the function value in other ways.

Support for ordering by a synthetic field is added precisely be able
sort by similarity.

pg_trgm has a few things **not** included here:

  - The notion of a similarity threshold for each of the three
similarity functions. Including the ability to show and set the
values.

  - Operators comparing a pair of values that each, implicitly use a
corresponding threshold resulting in a boolean.

  - Operators that compute "distance", defined as 1 - similarity, for
each of the three similarities.