Make Connections Thread Safe
Issuing multiple commands concurrently on the same connection results in
one of those commands hanging indefinitely, possibly due to a dead-lock
somewhere in libpq.

This introduces an MVar that is used to serialize access to the
connection for utilization purposes. The existing semantics around closing
connections should be preserved by the use of a second MVar.