Adding basic subquery expression support
This adds the ability to use a 'QueryExpr' as a subquery particularly
for one form of the expressions found at
https://www.postgresql.org/docs/12/functions-subquery.html.

In particular the following are added:
- 'EXISTS (subquery)', which results in a 'BooleanExpr'

- 'IN' subquery expression support of the form 'expression IN
(subquery)'.

- 'NOT IN' subquery expression support of the form 'expression NOT IN
(subquery)'.

- 'ANY' subquery expression support of the form 'expression operator
ANY (subquery)'. 'SOME' is an alias for 'ANY' supported by PostgreSQL,
but I left it out for simplicity.

- 'ALL' subquery expression support of the form 'expression operator
ALL (subquery)'.

Notably for 'IN', 'NOT IN', 'ANY', and 'ALL' there is a form that
works with a "row_constructor". That is not implemented here.

Also, there is a choice with the 'IN' and 'NOT IN' support. This
resuses the existing generation of those keywords but exposes separate
functions for building the 'BooleanExpr'. We could instead expose the
pieces and have users build the 'BooleanExpr' or we could implement
them as specializations of the 'anySubquery' and 'allSubquery'. That
specialization would look something like: 'inSubquery = anySubquery
equalsOp' and 'notInSubquery = allSubquery notEqualsOp'. Note those
would not generate the same literal SQL bytes. The path chosen here
feels like the best trade-off for users of the api at the cost of some
extra function definitions.