Adds JoinExpr to model SQL JOIN
The pieces needed to represent @JOIN@ are now available at the 'expr' level.

- 'JoinType' represents the INNER/LEFT/RIGHT/LATERAL portion.

- 'JoinConstraint' represents the portion that would determine how to
limit the result set. Currently only a helper for 'ON expr' is added.
This is left general enough to support the 'USING(columnNames)' syntax
as well in a future exercise.

- 'JoinExpr' represent the above pieces put together so that they can
be easily folded together into a 'FromItemExpr'.

Perhaps the most controversial change included is renaming
'TableReferenceList' to 'FromItemExpr'. This change more closely
matches the name in the PostgreSQL documentation. Which hopefully aids
clarity as at least some users did not realize that the "Table" in
that name was _not_ necessarily a user defined table like a
'TableDefinition', but could be something like a @JOIN@ or a subquery.