Adds `MigrationLockOptions`
These new options allow the user to control how many times Orville
will try to acquire the migration lock as well as the delay between
each attempt.

Note that this changes the default delay from the previous 10 milliseconds
to a randomized delay between 100 and 125 milliseconds. The default
max attempts remains set to 25.