Add conditional expression support
Adds support for "conditional expressions" as seen at
https://www.postgresql.org/docs/13/functions-conditional.html

These are not typical SQL functions, in that they must not be quoted
the same way. As such these cannot use the pre-existing support for
functions.

The module name of 'ConditionalExpr' is a bit odd in that there is no
type of the same name. However grouping these items does seem natural
as the PostgreSQL documentation groups them.