Adding windowing function support
Adds Expr-level support for create a resuable window clause.
This takes the form of @WINDOW window_name AS (window_definition)@.

Creating a window_definition is separate precisely because it is
reused from the above window clause and in selection as the portion
that comes after the @OVER@ keyword.

Building on the ability to express a window definition, is a window
function call. These functions have some additional requirements:
  - They are only valid in select or order by contexts
  - The use of the OVER keyword and a window definition is required.

Specific helpers for the following windowing functions are included:
  - rank
  - rowNumber
  - dense_rank
  - percent_rank