Update ON CONFLICT support
This changes the support for building ON CONFLICT portions of an
INSERT in a few key ways:

- Supports using an 'IndexBodyExpr' in addition to 'ConstraintName'
and 'Qualified ColumnName' for determining what to check for a
conflict.

- Is based more around 'RawSql', which allows for callers to inject
things not explicitly supported at this time. For example, we do not
have a notion of COLLATION, and the PostgreSQL documentation even
mentions it as usually unneeded. However this now allows for it.

- The support is also broken out to a separate module. This is more
about module size than anything else. For example, it is not likely
that we would reuse the support elsewhere. This implementation detail
is easy enough to change and should be reconsidered before the next
release.