Adds `ValuesExpr` for constructing `VALUES` SQL
I added a new module for constructing `VALUES` statements and refactored
`InsertSource` to use it. I deprecated the old `InsertSource`
construction functions, which could be used to produce invalid SQL by
passing empty lists of values.

This also had the effect of improving the behavior of `mkInsertSource`
when called with a `SqlMarshaller` with no fields; it will now use
`DEFAULT` for that case instead of generating invalid SQL.