Reduce the load in CI
After some recent failures regarding pull limits with dockerhub this
reduces the test matrix to only the "standard" stack.yaml for the
"middle" PostgreSQL versions.

PostgreSQL generally has 5 supported versions at a time. The idea
here is that if we work on all the GHC versions we support on the
oldest and newest of those PostgreSQL versions then a more minimal
check for the middle 3 is not risky. This cuts our CI build matrix in
half so we shouldn't hit dockerhub pull limits nearly as quickly.