Supports negative years for date and timestamp types
Previously we were ISO8601 encoding/decoding dates and timestamps with
negative years, however PostgreSQL uses a "BC" suffix to indicate
negative years. I added support for this syntax to the `Day`, `UTCTime`,
and `LocalTime` `SqlType`s, and expanded the range of years generated in
testing to ensure we're handling the negative years correctly.