Improves transaction rollback handling
I added exception handling around the success event SQL execution
in `finishTransaction` in `withTransaction`. If a `SqlExecutionError`
occurs, it will run the rollback callback and rethrow the exception.

I also added libpq transaction status checks to `beginTransaction` and
`finishTransaction` in `withTransaction`. If the status is unexpected
for a given transaction event, `withTransaction` will throw an
`UnexpectedTransactionStatusError`.