Improves transaction rollback handling
I added a libpq transaction status check to the `BracketSuccess` branch
of `withTransaction` that will ensure that the transaction has the
`TransInTrans` state before issuing the success callback. If the status
is `TransInError`, the transaction will be rolled back. If it is
anything else, the transaction will be rolled back and an
`UnexpectedTransactionStatusError` will be thrown.

Additionally, I added exception handling around the success SQL
execution that will run the rollback callback if executing the SQL fails
with a `SqlExecutionError`, which could happen if a transaction cannot
be committed due to something like a serialization error.