Add initial support for identity columns
- Expr support for column identity is added for columns that should be
generated always or only by default

  - There is some oddity here. The specification for creating an
identity column is to add a constraint. However when looking up in the
catalog identity information is not held with other constraints.

- Field definitions can be marked as identity

- Support for identity at the field definition limits to non-nullable
fields, as is the case in PostgreSQL.

- AutoMigration support for marking column identity

There is additional work needed for full support. This is left for the
future. Particularly, identity columns can take options for the
underlying sequence. However, supporting that will require being able
to look up said sequence using a catalog table not currently
supported: pg_depend.