Rolls back transactions when begin fails in `withTransaction`
Previously Orville could leak a connection with an open transaction if
the `BeginTransaction` callback failed, e.g. due to throwing an
exception or being interrupted by an async exception. We added exception
handling around the begin action and added a test that will fail if the
exception is not handled correctly.

Co-authored-by: Ram <ram@flipstone.com>