Wraps `withTransaction` cleanup in `uninterruptibleMask`
`withTransaction` can perform blocking/interruptible IO when cleaning up
a transaction after an exception or when attempting to commit the
transaction, so previously it was possible for an async exception to
interrupt the cleanup and prevent the transaction from being committed
properly.

I wrapped `finishTransaction` in `uninterruptibleMask` so that this
cannot happen. The transaction callbacks are restored to the outer
masking state to reduce the chances of a callback blocking forever
without being interruptible.

I also added a test that will fail if the `uninterruptibleMask` is
removed.