Refactor Constraints to support named constraint migration
This refactors constraint handling to allow for arbitrary constraint
expressions that are then auto-migrated based solely on their name. The
initial inspiration was allowing the flexibility for a user to create a
constraint like "UNIQUE NULLS NOT DISTINCT" which is not currently
supported in orville. This change also subsumes check constraint
handling into name based constraints, as they were already migrated
according to name only. As a nice side effect the constraint migration
handling code has been refactored to be slightly simpler as well.