Fix date generation range and haddock comments
Because the '0' turns into 1 BCE, as we correctly handle with a
subtraction for negative numbers, and PostgreSQL having a lower bound
for dates of 4713 BCE, there was an off by 1 error in our date
generator for tests. Namely that '1 - (-4713)' is 4714, falling off
the allowed minimum in PostgreSQL. This adjusts the generator to work
with a lower bound of -4712.

While here, I noticed a typo in our haddocks that had -4731 instead of
-4713, but that too would have resulted in the off by one so the
comments have been adjusted.

I would think this has never been hit in practice as storing dates
that far back is probably rare for users of Orville.